requirements.txt
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.1.0
jupyter-core==4.6.3
matlab
# matlabengineforpython===R2020a
matplotlib
notebook
numpy
opencv-python
Pillow
PyYAML
scikit-image
scipy
tqdm